FastAPI_Server/requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
transformers
sentence-transformers
pytz
ipython
pillow
langgraph
tavily-python
langchain-google-genai
langchain_tavily
trafilatura
boto3
unstract
llmwhisperer-client
faiss-cpu
langchain_ollama
mem0ai
langgraph-checkpoint-redis
fastapi
uvicorn
redis
google-cloud-storage
python-multipart
pydantic
python-dotenv
gunicorn
langgraph[sqlite]
aiosqlite
